Rename files that were in camelCase.
