+++ components/receive-amount.tsx
