+++ components/receive-no-amount.tsx
